LPS-101372 Move to util class
