+++ modules/sdk/gradle-plugins-defaults/src/main/java/com/liferay/gradle/plugins/defaults/internal/MavenDefaultsPlugin.java
@@ -19,6 +19,7 @@ import com.liferay.gradle.plugins.defaults.LiferayOSGiDefaultsPlugin;
 import com.liferay.gradle.plugins.defaults.internal.util.FileUtil;
 import com.liferay.gradle.plugins.defaults.internal.util.GradlePluginsDefaultsUtil;
 import com.liferay.gradle.plugins.defaults.internal.util.GradleUtil;
+import com.liferay.gradle.plugins.defaults.internal.util.LiferayRelengUtil;
 
 import java.io.File;
 
@@ -103,8 +104,8 @@ public class MavenDefaultsPlugin extends BaseDefaultsPlugin<MavenPlugin> {
 					return;
 				}
 
-				File relengDir = LiferayRelengPlugin.getRelengDir(project);
-				File releaseRelengDir = LiferayRelengPlugin.getRelengDir(
+				File relengDir = LiferayRelengUtil.getRelengDir(project);
+				File releaseRelengDir = LiferayRelengUtil.getRelengDir(
 					releaseProjectDir);
 
 				if ((relengDir == null) && (releaseRelengDir != null) &&
